+++ lib/src/di/app_dependencies.dart
@@ -15,6 +15,8 @@ import 'package:where_to_go_today/src/features/authservices/repository/auth_repo
 import 'package:where_to_go_today/src/features/main/places/service/places_bloc.dart';
 import 'package:where_to_go_today/src/features/main/places/service/repository/places_repository.dart';
 import 'package:where_to_go_today/src/features/onboard/services/onboarding_bloc.dart';
+import 'package:where_to_go_today/src/features/onboard/services/repository/onboard_repository.dart';
+import 'package:where_to_go_today/src/features/onboard/services/storage/onboard_storage.dart';
 import 'package:where_to_go_today/src/features/settings/service/event/settings_event.dart';
 import 'package:where_to_go_today/src/features/settings/service/repository/settings_repository.dart';
 import 'package:where_to_go_today/src/features/settings/service/settings_bloc.dart';
@@ -25,11 +27,13 @@ class AppDependencies extends DependencyBundle {
   final dio = DioModule().dio;
   final settingsController = SettingsBloc(SettingsRepository());
   final tokenStorage = TokenStorage();
+  final onboardStorage = OnboardStorage();
   final facebookAuthService = FacebookAuthService();
 
   final googleAuth = GoogleAuth();
   late final vkAuth = VKAuth();
 
+  late final onboardRepository = OnboardRepository(onboardStorage);
   late final authRepository = AuthRepository(AuthApi(dio));
   late final onboardingBloc = OnboardingBloc();
   late final authBloc = AuthBloc(
@@ -55,6 +59,7 @@ class AppDependencies extends DependencyBundle {
     settingsController.add(LoadSettings());
 
     await tokenStorage.init();
+    await onboardStorage.init();
 
     if (Firebase.apps.isNotEmpty) {
       await Firebase.initializeApp(