+++ lib/src/res/asset.dart
@@ -15,12 +15,16 @@ class _Png {
 
 class _Svg {
   static const _pathSignIn = 'assets/images/auth/sign_in/';
-  static const _pathOnboarding = 'assets/images/onboarding/';
+  static const _pathProfile = 'assets/images/profile/';
   static const _pathPlace = 'assets/images/place/';
+  static const _pathOnboarding = 'assets/images/onboarding/';
 
   final iconFacebook = '${_pathSignIn}icon_facebook.svg';
   final iconGoogle = '${_pathSignIn}icon_google.svg';
   final iconVkontakte = '${_pathSignIn}icon_vkontakte.svg';
+  final iconEdit = '${_pathProfile}edit.svg';
+  final iconSettings = '${_pathProfile}settings.svg';
+  final iconCocktail = '${_pathPlace}icon_cocktail.svg';
 
   final onboardingFirst = '${_pathOnboarding}onboarding_1.svg';
   final onboardingSecond = '${_pathOnboarding}onboarding_2.svg';