+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - feat: SignInScreenVm - add phone authorization
 - feat: OnboardingScreen
 - feat: MainScreen + PlacesScreen
+- feat: OnboardStorage
+- feat: profile screen
 
 ## 0.3.0